+++ src/main/java/com/baichen/rpc/api/User.java
@@ -1,0 +1,9 @@
+package com.baichen.rpc.api;
+
+import lombok.Data;
+
+@Data
+public class User {
+    private String name;
+    private int age;
+}