+++ src/main/java/com/baichen/rpc/consumer/GenericConsumer.java
@@ -1,0 +1,6 @@
+package com.baichen.rpc.consumer;
+
+public interface GenericConsumer {
+
+    Object $invoke(String serviceName, String methodName, String[] paramTypes, Object[] args);
+}